Added the comments feature

+++ backend/src/controllers/post.controllers.js
@@ -3,6 +3,8 @@ const { toFile } = require('@imagekit/nodejs');
 const sharp = require("sharp")
 const postModel = require("../models/posts.model");
 const likeModel = require('../models/likes.model');
+const userModel = require("../models/users.model");
+const commentModel = require('../models/comments.model');
 
 async function createPostController(req, res) {
     const { caption } = req.body
@@ -34,23 +36,36 @@ async function createPostController(req, res) {
 
 async function likePostController(req, res) {
     let user = req.user.id;
-    let post = req.params.id
+    let post = req.params.id;
 
-    const alreadyLiked = await likeModel.findOne({
-        user,
-        post
-    })
+    const [userExists, postExists, alreadyLiked] = await Promise.all([
+        userModel.exists({ _id: user }),
+        postModel.exists({ _id: post }),
+        likeModel.exists({ user, post })
+    ]);
 
-    if(alreadyLiked) {
+    if (!userExists) {
+        return res.status(404).json({
+            message: "User not found"
+        })
+    }
+
+    if (!postExists) {
+        return res.status(404).json({
+            message: "Post not found"
+        })
+    }
+
+    if (alreadyLiked) {
         return res.status(409).json({
             message: "Post already liked"
         })
     }
 
-    const like = await likeModel.create({
+    await likeModel.create({
         user,
         post
-    }) 
+    })
 
     res.status(201).json({
         message: "Post liked successfully"
@@ -61,10 +76,34 @@ async function unlikePostController(req, res) {
     let user = req.user.id;
     let post = req.params.id
 
+    const [userExists, postExists, notLiked] = await Promise.all([
+        userModel.exists({ _id: user }),
+        postModel.exists({ _id: post }),
+        likeModel.exists({ user, post })
+    ]);
+
+    if (!userExists) {
+        return res.status(404).json({
+            message: "User not found"
+        })
+    }
+
+    if (!postExists) {
+        return res.status(404).json({
+            message: "Post not found"
+        })
+    }
+
+    if (!notLiked) {
+        return res.status(409).json({
+            message: "User haven't liked the post"
+        })
+    }
+
     await likeModel.deleteOne({
         user,
         post
-    }) 
+    })
 
     res.status(200).json({
         message: "Post unliked successfully"
@@ -72,6 +111,16 @@ async function unlikePostController(req, res) {
 }
 
 async function getFeedController(req, res) {
+    const user = req.user.id
+
+    const userExists = await userModel.exists({ _id: user })
+
+    if (!userExists) {
+        return res.status(404).json({
+            message: "User not found"
+        })
+    }
+
     const posts = await postModel.find()
 
     res.status(200).json({
@@ -80,6 +129,44 @@ async function getFeedController(req, res) {
     })
 }
 
+async function commentController(req, res) {
+    const user = req.user.id
+    const post = req.params.id
+    const { comment } = req.body
+
+    const [userExists, postExists] = await Promise.all([
+        userModel.exists({ _id: user }),
+        postModel.exists({ _id: post })
+    ]);
+
+    if (!userExists) {
+        return res.status(404).json({
+            message: "User not found"
+        })
+    }
+
+    if (!postExists) {
+        return res.status(404).json({
+            message: "Post not found"
+        })
+    }
+
+    if (!comment || !comment.trim()) {
+        return res.status(400).json({ message: "Comment cannot be empty" });
+    }
+
+    const newComment = await commentModel.create({
+        comment,
+        post,
+        user
+    })
+
+    res.status(201).json({
+        message: "User commented successfully",
+        newComment
+    })
+}
+
 module.exports = {
-    createPostController, getFeedController, likePostController, unlikePostController
+    createPostController, getFeedController, likePostController, unlikePostController, commentController
 }
